Make IPAAbstractVersion available to all platforms
They still need to implement own comparison function
